Refactor StudyStatistics Page
